generator/requirements.txt
psycopg2-binary==2.9.9
Faker==20.0.0